+++ tools/gulp/clear.js
