+++ tools/gulp/copyAssets.js
