+++ tools/gulp/copyFonts.js
