+++ tools/gulp/ensureAppIsBuilt.js
